+++ R/attackerkb-package.R
@@ -4,6 +4,9 @@
 #' various aspects of emergent or existing vulnerabilities and the community can query and
 #' retrieve results. Tools are provided to query the AttackerKB API.
 #'
+#' For all API calls, set the option `progress_enabled` to `FALSE` to disable
+#' progress spinner.
+#'
 #' @md
 #' @name attackerkb
 #' @keywords internal
@@ -11,4 +14,5 @@
 #' @import httr
 #' @importFrom jsonlite fromJSON
 #' @importFrom data.table rbindlist
+#' @importFrom progress progress_bar
 "_PACKAGE"