+++ NEWS.md
@@ -1,2 +1,6 @@
+0.1.1
+* Fixed an issue where errors wld be thrown if zero API results came back
+* Added progress spinner for API calls vs the hacky dots
+
 0.1.0 
 * Initial release